Token in headers, let's go!

+++ API/controllers/commentController.js
@@ -1,18 +1,38 @@
 let Comment = require("../models/commentModel");
+let jwt = require("../utilities/jwt");
 
 module.exports.deleteComment = (req, res) => {
-    Comment.findOneAndDelete({ _id: req.params.comment_id })
-        .exec(function (err, comment) {
+    Comment.findOne({ _id: req.params.comment_id })
+        .exec((err, comment) => {
             if (err) {
                 res.json({
-                    status: "Couldn't delete this comment!"
+                    status: "Trouble finding this comment!"
                 })
             }
             else {
-                res.json({
-                    status: "Comment deleted successfully!",
-                    data: comment
-                })
+                let auth_token = jwt.verify(req.headers.authentication);
+                if (auth_token.uid == comment.creator) {
+                    Comment.findOneAndDelete({ _id: req.params.comment_id })
+                        .exec(function (err, comment) {
+                            if (err) {
+                                res.json({
+                                    status: "Couldn't delete this comment!"
+                                })
+                            }
+                            else {
+                                res.json({
+                                    status: "Comment deleted successfully!",
+                                    data: comment
+                                })
+                            }
+                        })
+                }
+                else {
+                    res.json({
+                        status: "This is not your comment!"
+                    })
+                }
             }
         })
+
 }
